spouts: add getSourceIcon() method
To allow distinguishing between an icon coming from the feed item and one from the feed itself,
we added a new optional getSourceIcon() method. When the getIcon() method returns null for an item
or the item icon cannot be fetched, the source icon will be used as a fallback.

Both getSourceIcon() and getIcon() can return null now, if an icon is not available.